services/backend/pyproject.toml
[tool.poetry]
name = "kurisu-backend"
version = "1.0.0"
description = "The backend service for the Kurisu project."
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
uvicorn = { extras = ["standard"], version = "^0.35.0" }
matplotlib = "^3.10.6"
pillow = "^11.3.0"
opencv-python-headless = "^4.12.0"
wand = "^0.6.13"
python-magic = "^0.4.27"
pyfiglet = "^1.0.4"
motor = "^3.7.1"
pymongo = "^4.14.1"
redis = "^6.4.0"
celery = "^5.5.3"
structlog = "^25.4.0"
httpx = {extras = ["http2"], version = "^0.28.1"}
openai = "^1.106.1"
fal-client = "^0.7.0"
imgkit = "^1.2.3"
jinja2 = "^3.1.6"
pytz = "^2025.2"
kurisu-core = { path = "../../packages/kurisu_core", develop = true }
prometheus-fastapi-instrumentator = "^7.1.0"
opentelemetry-api = "^1.36.0"
opentelemetry-sdk = "^1.36.0"
opentelemetry-instrumentation-fastapi = "^0.57b0"
opentelemetry-exporter-otlp-proto-grpc = "^1.36.0"
opentelemetry-instrumentation-httpx = "^0.57b0"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.1.0"
pytest-dotenv = "^0.5.2"
ruff = "^0.12.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["../tests/backend"]
pythonpath = ["."]
asyncio_mode = "auto"
env_files = ["../.env.test"]